Export features for vocabularies

+++ public/javascripts/application.js
@@ -114,19 +114,21 @@ function paintNewGraph(scores, offset) {
 // Redirects to given action
 
 function redirectTo(url, text) {
-	if (text.substring(0,5) == 'Apply') {
-		if (confirm('Are you sure you want to copy these attributes to all translations?')) {
-			new Ajax.Request(url, { method: 'post', onSuccess: function() { highlight(); } });
+	if (url != "") { 
+		if (text.substring(0,5) == 'Apply') {
+			if (confirm('Are you sure you want to copy these attributes to all translations?')) {
+				new Ajax.Request(url, { method: 'post', onSuccess: function() { highlight(); } });
+			}
+			$('redirect_to').selectedIndex = 0;
 		}
-		$('redirect_to').selectedIndex = 0;
-	}
-	else if (text.substring(0,6) == 'Delete') {
-		if (confirm('Are you sure you want to delete his vocabulary?')) {
-			new Ajax.Request(url, { method: 'delete', onSuccess: function() { window.location = '/vocabularies' } });
+		else if (text.substring(0,6) == 'Delete') {
+			if (confirm('Are you sure you want to delete his vocabulary?')) {
+				new Ajax.Request(url, { method: 'delete', onSuccess: function() { window.location = '/vocabularies' } });
+			}
+			$('redirect_to').selectedIndex = 0;
 		}
-		$('redirect_to').selectedIndex = 0;
-	}
-	else {
-		window.location = url;
+		else {
+			window.location = url;
+		}
 	}
 }
